I have started building upon a new word-based game. Currently, I have 3 classes excluding Entity, which the parent class for entities: Warrior, Medic and archer which will contain the attributes of entities with each one having their unique attributes and abilities.
I have not added unique attributes yet, meaning that Warrior, Medic and Archer are all identical right now. This is the beginning phase of the word game. I have used user input to determine which class they would prefer and what they would like to be called, that is all. More development will begin soon.

Please ignore the methods and constructors in other classes as they are not being used currently...